requirements.txt
asgiref==3.6.0
astroid==2.13.2
async-timeout==4.0.2
autopep8==2.0.1
bcrypt==4.0.1
boto==2.49.0
boto3==1.26.88
botocore==1.29.88
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
dill==0.3.6
Django==3.2
django-crispy-forms==1.14.0
django-excel==0.0.10
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
fabric==3.0.0
Flask==2.2.2
future==0.18.3
idna==3.4
invoke==2.0.0
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.9.0
lml==0.1.0
MarkupSafe==2.1.2
mccabe==0.7.0
MouseInfo==0.1.3
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
paramiko==3.0.0
Pillow==9.4.0
platformdirs==2.6.2
prettytable==3.6.0
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyGetWindow==0.0.9
pylint==2.15.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyodbc==4.0.35
PyPDF2==3.0.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.7
redis==4.5.5
reportlab==3.6.13
requests==2.28.2
s3transfer==0.6.0
serializers==0.2.4
shall==0.1
shell==1.0.1
six==1.16.0
speedtest==0.0.1
sqlparse==0.4.3
texttable==1.6.7
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.7
xlwt==1.3.0
